Clean code interface
