+++ crates/memory/src/interface.rs
@@ -1,58 +1,14 @@
-use super::mbc::{Cartridge, Mbc0, MbcBus};
 use super::wram;
 use crate::bios;
+use crate::mbc::Mbc;
 use crate::memory;
-use crate::state::State;
 use std::cell::RefCell;
-use std::convert::From;
 use std::rc::Rc;
 
 pub type Memory = Rc<RefCell<memory::Memory>>;
 
-pub trait NewMemory {
-    fn new(mbc: Cartridge, data: Vec<u8>) -> Self;
-}
-
-struct Data(State, Bios, Rom, Wram);
-
-impl From<Data> for memory::Memory {
-    fn from(data: Data) -> Self {
-        Self {
-            state: data.0,
-            bios: data.1,
-            rom: data.2,
-            wram: data.3,
-        }
-    }
-}
-
-impl NewMemory for Memory {
-    fn new(mbc: Cartridge, data: Vec<u8>) -> Self {
-        let rom: Rom = Rc::new(RefCell::new(match mbc {
-            Cartridge::Mbc0 => Mbc0::new(data),
-            _ => unimplemented!(),
-        }));
-        let state = State::Bios;
-        let bios = Bios::default();
-        let wram = Wram::default();
-        let memory: memory::Memory = Data(state, bios, rom, wram).into();
-
-        Rc::new(RefCell::new(memory))
-    }
-}
-
 pub type Wram = Rc<RefCell<wram::Wram>>;
 
-pub type Rom = Rc<RefCell<dyn MbcBus>>;
-
-pub trait RomDefault {
-    fn default() -> Self;
-}
-
-impl RomDefault for Rom {
-    fn default() -> Self {
-        Rc::new(RefCell::new(Mbc0::default()))
-    }
-}
+pub type Rom = Rc<RefCell<Box<dyn Mbc>>>;
 
 pub type Bios = Rc<RefCell<bios::Bios>>;
